Project-wide refactoring
- SafeAreaPadding and SafeAreaAdaptation has "Area" removed because it was reading like we "pad the safe area" instead of padding using the safe area. Plus in the future if these component may account for cutouts in addition to safe area, then just "safe" make broader sense. (Because not hitting the cutouts is also safe) Also this prepares for a new component in the same series which is called "SafePositioning" which would be more awkward when called "SafeAreaPositioning".
- Utility methods that is in the main assembly but intended for only editor use is separated to a new file with a new class name `NotchSolutionUtilityEditor`.
- Enums renamed so they also make sense in the new WIP  `SafePositioning` component.
- Things which should be `internal` made `internal`.
- New base class for `SafePadding` in preparation for `SafePositioning` to also inherits from.
