+++ .Documentation/Components.md
@@ -2,7 +2,7 @@
 
 # Components
 
-* <img src="../Icons/SafeAreaPaddingIcon.png" width="18"> [SafeAreaPadding](Components/SafeAreaPadding.md): `RectTransform` layout component for uGUI based on the value returned by [`Screen.safeArea`](https://docs.unity3d.com/ScriptReference/Screen-safeArea.html).
+* <img src="../Icons/SafeAreaPaddingIcon.png" width="18"> [SafePadding](Components/SafePadding.md): `RectTransform` layout component for uGUI based on the value returned by [`Screen.safeArea`](https://docs.unity3d.com/ScriptReference/Screen-safeArea.html).
 * [Adaptation Components](Components/AdaptationComponents.md): Adapts any game object with Animation Playables API and `Animator` component.
   * <img src="../Icons/AspectRatioAdaptationIcon.png" width="18"> [AspectRatioAdaptation](Components/AdaptationComponents.md#-aspectratioadaptation): Dynamically changes anything based on the ratio of the screen.
   * <img src="../Icons/SafeAreaAdaptationIcon.png" width="18"> [SafeAreaAdaptation](Components/AdaptationComponents.md#-safeareaadaptation): Dynamically changes anything based on the safe area.
