+++ Editor/AspectRatioAdaptationDrawer.cs
@@ -1,0 +1,7 @@
+using UnityEditor;
+
+namespace E7.NotchSolution
+{
+    [CustomEditor(typeof(AspectRatioAdaptation))]
+    public class AspectRatioAdaptationDrawer : AdaptationBaseDrawer { }
+}
