+++ Editor/NoSo-MockupCanvas.prefab
@@ -69,7 +69,6 @@ MonoBehaviour:
   m_GameObject: {fileID: 3673283291737537849}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_GeneratorAsset: {fileID: 0}
   m_Script: {fileID: 1980459831, guid: f70555f144d8491a825f0804e09c671c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
@@ -92,7 +91,6 @@ MonoBehaviour:
   m_GameObject: {fileID: 3673283291737537849}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_GeneratorAsset: {fileID: 0}
   m_Script: {fileID: 11500000, guid: 83b8ec99fbf59473891f0f467ac098e7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
@@ -149,9 +147,8 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 3673283292801682466}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
-  m_GeneratorAsset: {fileID: 0}
   m_Script: {fileID: -765806418, guid: f70555f144d8491a825f0804e09c671c, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
