+++ Editor/SafeAdaptationDrawer.cs.meta
