+++ Runtime/Base/Adaptation/PerEdgeAdaptations.cs
@@ -3,5 +3,5 @@
 namespace E7.NotchSolution
 {
     [Serializable]
-    public class PerEdgeAdaptations : PerEdgeValues<BlendedClipsAdaptor> { }
+    internal class PerEdgeAdaptations : PerEdgeValues<BlendedClipsAdaptor> { }
 }
