+++ Runtime/Base/EdgeEvaluationMode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 33767150048334d38bca1d394fa48b3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
