+++ Runtime/Base/InternalsVisibleTo.cs
@@ -1,0 +1,2 @@
+using System.Runtime.CompilerServices;
+[assembly: InternalsVisibleTo("E7.NotchSolution.Editor")]
