+++ Runtime/Base/InternalsVisibleTo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb43dfddc128d42b7adb47948e4d9b90
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
