+++ Runtime/Base/NotchSolutionUIBehaviourBase.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9dc77cbf56135463fb57febb35d77297
+guid: cb167ee7953de4241ac2a682af593c8d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
