+++ Runtime/Base/PerEdgeEvaluationModes.cs
@@ -3,5 +3,5 @@ using System;
 namespace E7.NotchSolution
 {
     [Serializable]
-    public class PerEdgeEvaluationModes : PerEdgeValues<SafeAreaEvaluationMode> { }
+    internal class PerEdgeEvaluationModes : PerEdgeValues<EdgeEvaluationMode> { }
 }
