+++ Runtime/Base/PerEdgeValues.cs
@@ -4,7 +4,7 @@ using System;
 namespace E7.NotchSolution
 {
     [Serializable]
-    public class PerEdgeValues<T>
+    internal class PerEdgeValues<T>
     {
         [SerializeField] public T left;
         [SerializeField] public T bottom;
