+++ Runtime/Components/SafeAdaptation.cs.meta
