+++ package.json
@@ -2,7 +2,7 @@
     "name": "com.e7.notch-solution",
     "author": "Sirawat Pitaksarit / 5argon - Exceed7 Experiments",
     "displayName": "Notch Solution",
-    "version": "1.6.0",
+    "version": "1.7.0",
     "unity": "2019.1",
     "keywords": [
         "UI",
